chore: Rename SwipeToDismissComponent and move it to correct package
